renaming and reorganizing components
