+++ src/components/Candidates/Card/Card.css
